Allow more customization on sink output, fix some errors

+++ src/OpenLoco/src/Ui/Dropdown.cpp
@@ -99,7 +99,7 @@ namespace OpenLoco::Ui::Dropdown
                 }
 
                 default:
-                    Logging::error("Unknown format: {}", arg.type);
+                    Logging::error("Unknown format: {}", static_cast<int>(arg.type));
                     break;
             }
         }
